Fix local-label computation for expressions, statements, decls, and specs.
Previously, we computed the labels before checking whether the expression/statement/decl/spec is `nil`, which would cause us to spuriously allocate labels for `nil` nodes. These labels weren't actually used subsequently, so in practice this was relatively harmless.

I've also added a few more missing `nil` guards. Some of them are possibly superfluous due to implicit constraints on the shape of the AST, but I'd rather not rely on that.